2004-11-20      Thomas Doerfler <Thomas.Doerfler@imd-systems.de>
        PR 703/filesystem
	* ide/pcmcia_ide.c: Extend the "modes_available" and "mode_active" parameters
	to uint16_t for UDMA support.

+++ c/src/lib/libbsp/powerpc/mbx8xx/ide/pcmcia_ide.c
@@ -356,8 +356,8 @@ rtems_status_code mbx8xx_pcmciaide_config_io_speed
 +---------------------------------------------------------------------------+
 | Input Parameters:                                                         |
 \*-------------------------------------------------------------------------*/
- int       minor,                   /* controller minor number       */
- uint8_t   modes_avail              /* optional argument             */
+ int        minor,                   /* controller minor number       */
+ uint16_t   modes_avail              /* optional argument             */
  )
 /*-------------------------------------------------------------------------*\
 | Return Value:                                                             |
